requirements.txt
#streamlit==1.31.0
#crewai==0.1.32
#langchain-groq==0.0.1
streamlit
crewai
langchain-groq
chromadb
pysqlite3-binary
langchain-core
langchain-openai
langchain-community
load_dotenv
unstructured
pyowm
tools
wikipedia
yfinance
sec-api
tiktoken
faiss-cpu
python-dotenv
agentops
weave
